+++ backend_restaurantmanager/src/main/java/com/tfg_rm/backend_restaurantmanager/websocket/interceptor/JwtHandshakeInterceptor.java
@@ -59,8 +59,8 @@ public class JwtHandshakeInterceptor implements HandshakeInterceptor {
 
                 // Validate the token and extract user information if valid
                 if (jwtService.validateToken(token)) {
-                    Integer restaurantId = jwtService.getRestaurantId(token);
-                    Integer userId = jwtService.getUserId(token);
+                    Long restaurantId = jwtService.getRestaurantId(token);
+                    Long userId = jwtService.getUserId(token);
                     String role = jwtService.getRole(token);
 
                     attributes.put("restaurantId", restaurantId);
